requirements.txt
pyzmq>=14.3.1
crypto>=1.1.0
tornado>=4.1
sqlalchemy>=0.9.7
requests>=2.3.0
appdirs
mailchimp
pycountry
pyboleto
mock
